backend/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.22
pandas>=2.0
networkx>=3.0
python-dotenv>=1.0
pydantic>=2.5
google-generativeai>=0.8.0
httpx>=0.27
